fix(lark): 修复 Codex review 三点——Mir 折叠正文/分隔符 locale/hint 文案
回应 PR #400 review：

1. Mir 看不到折叠正文（回归）：`summarizeAvailableBots` 只读 `hint` 和 `<bot/>`
   子节点，>3 折叠后名字在 tag body 里被丢弃。把纯 prompt 归一化逻辑从
   `mir-runner.ts`（import 即跑 main()，不可测）抽到新的 side-effect-free
   `src/mir-prompt.ts`，并在 `summarizeAvailableBots` 里 decode 并追加非空
   正文（内联变体剥掉 `<bot/>` 后为空，no-op）；顺手把硬编码的「handoff」
   窄口径改成 communicate/collaborate。

2. 折叠分隔符与 t() 的 locale 解析不一致：原来按裸 `locale` 判断，
   `setDefaultLocale('en')` 且调用方没传 locale 时会输出英文句子却用中文顿号。
   改为 `locale ?? getDefaultLocale()` 解析 effective locale 再选分隔符。

3. `identity.mention_partners` 文案过期：仍说 open_id 列在「每条用户消息」里，
   但新逻辑是首轮才有、>3 不列 open_id。改成「首轮上下文提示可协作 bot
   （少含 open_id、多只列名字）；open_id 也可 botmux bots list 查」。zh+en 同步。

测试：新增 `test/available-bots-collapse.test.ts`（7 项）——内联/折叠/count、
分隔符按 default locale、Mir 两种形态均 surface 名字、空块缺省。
本文件 + i18n(487) + available-bots + session-manager-scan + mir 邻近套件
= 524 tests 全绿；`pnpm build` 绿。
