+++ src/i18n/en.ts
@@ -534,7 +534,7 @@ export const messages: Record<string, string> = {
   'ai.identity.rule_no_proactive_pull': '- **Do not proactively pull other bots in by default.** Unless the user explicitly asks, or a segment can only be done by another bot, finish your part and stop.',
   'ai.identity.mention_intro': '**Hard fact about cross-bot collaboration**: in a Lark topic group, other bots **do NOT receive** the messages you send via `botmux send` by default —',
   'ai.identity.mention_must': 'To communicate or collaborate with another bot (so it receives your message), you **MUST** explicitly pass `--mention <other-bot-open-id>`. Without `--mention`, the other bot is not triggered at all.',
-  'ai.identity.mention_partners': '- Collaborator `open_id`s are listed in the `<available_bots>` block attached to every user message. You can also run `botmux bots list`.',
+  'ai.identity.mention_partners': '- The first-turn `<available_bots>` block lists the bots you can currently collaborate with (with open_ids when there are few, names only when many); you can also run `botmux bots list` anytime to get an open_id.',
   'ai.identity.mention_usage': '- Usage: `botmux send --mention ou_xxx "message"` (repeat `--mention` for multiple bots). `--mention-back` @s back whoever (person or bot) triggered this turn — open_id is pulled from the session, no need to type it.',
   'ai.identity.mention_gate': '- **@ hard-gate**: every `botmux send` MUST explicitly pick one or it errors out — `--mention` (name someone) / `--mention-back` (@ the triggerer) / `--no-mention` (none). Choose by message VALUE: a substantive conclusion the other party should read/confirm/decide on → --mention-back (or --mention); pure record / low-priority progress / short ack → --no-mention; a contentless "got it" is better not sent at all. Do not let --no-mention become the default, and do not @ people for nothing.',
   'ai.identity.mention_when_to': '- When to `--mention`: you need to communicate or collaborate with the other bot, the user explicitly asks for a handoff, you\'re passing a sub-task to it, or you need them to give the final result / take an independent action.',
